Unset 'own_window' option
With the 'own_window' option enabled placement of the widget is not
correct: instead of the intended position it is just placed at the upper
left corner of the screen. At the intended position a black rectangle is
visible instead.
Disabling 'own_window' changes the behavior to correctly display the
widget.