Display previous and following months
Until now we were displaying the current month only. In some instances,
however, one needs to check a date one month in the past or one in the
future.
With this change we display a calendar covering the previous, the
current, and the next month. The calendars are stacked on top of each
other.