Correct wrong previous/next month being displayed
On 2016/08/31 it was witnessed that the next month being displayed was
October and not September. The problem is believed to be date's
wrong/misleading next month calculation.
This change works around this issue by calculating the next/previous
month based on the 15th of the current month. By doing so, we avoid any
issues with unusually short/long months when performing simple month
arithemtics.